CalEnviroScreen 4.0 Data

A dataset used in the DaSEH Short Course. The CalEnviroScreen object contains data from CalEnviroScreen 4.0, including the environmental dataset and the demographic dataset.

calenviroscreen

Format

A data frame with 8035 rows and 66 columns.

CensusTract

Census Tract ID from 2010 Census

CaliforniaCounty

California county that the census tract falls within

ZIP

Postal ZIP Code that the census tract falls within

Longitude

Longitude of the centroid of the census tract

Latitude

Latitude of the centroid of the census tract

ApproxLocation

Approximate city, town, or area where each census tract is located based on US Census Incorporated Places (2020, Cities), US Census Designated Places, (2020, Designated Places), and the CA Department of Tax and Fees City and County Boundaries and City Annexations (2021, Unincorporated Areas) boundary files. All tracts that did not fall within one of these boundaries were assigned "unincorporated county area" based on that tract's county. This is for reference purposes only and should not be used to determine whether a census tract falls within a city or town boundary.

CES4.0Score

CalEnviroScreen Score, Pollution Score multiplied by Population Characteristics Score

CES4.0Percentile

Percentile of the CalEnviroScreen score

CES4.0PercRange

Percentile of the CalEnviroScreen score, grouped by 5 percent increments

Ozone

Amount of daily maximum 8 hour Ozone concentration

OzonePctl

Ozone percentile

PM2.5

Annual mean PM2.5 concentrations

PM2.5.Pctl

PM2.5 percentile

DieselPM

Diesel PM emissions from on-road and non-road sources

DieselPMPctl

Diesel PM percentile

DrinkingWater

Drinking water contaminant index for selected contaminants

DrinkingWaterPctl

Drinking water percentile

Lead

Potential risk for lead exposure in children living in low-income communities with older housing

LeadPctl

Children's lead risk from housing percentile

Pesticides

Total pounds of selected active pesticide ingredients (filtered for hazard and volatility) used in production-agriculture per square mile

PesticidesPctl

Pesticides percentile

ToxRelease

Toxicity-weighted concentrations of modeled chemical releases to air from facility emissions and off-site incineration (from RSEI)

ToxReleasePctl

Toxic release percentile

Traffic

Traffic density in vehicle-kilometers per hour per road length, within 150 meters of the census tract boundary

TrafficPctl

Traffic percentile

CleanupSites

Sum of weighted EnviroStor cleanup sites within buffered distances to populated blocks of census tracts

CleanupSitesPctl

Cleanup sites percentile

GroundwaterThreats

Sum of weighted GeoTracker leaking underground storage tank sites within buffered distances to populated blocks of census tracts

GroundwaterThreatsPctl

Groundwater threats percentile

HazWaste

Sum of weighted hazardous waste facilities and large quantity generators within buffered distances to populated blocks of census tracts

HazWastePctl

Hazardous waste percentile

ImpWaterBodies

Sum of number of pollutants across all impaired water bodies within buffered distances to populated blocks of census tracts

ImpWaterBodiesPctl

Impaired water bodies percentile

SolidWaste

Sum of weighted solid waste sites and facilities (SWIS) within buffered distances to populated blocks of census tracts

SolidWastePctl

Solid waste percentile

PollutionBurden

Average of percentiles from the Pollution Burden indicators (with a half weighting for the Environmental Effects indicators)

PollutionBurdenScore

Pollution Burden variable scaled with a range of 0-10. (Used to calculate CES 4.0 Score)

PollutionBurdenPctl

Pollution burden percentile

Asthma

Age-adjusted rate of emergency department visits for asthma

AsthmaPctl

Asthma percentile

LowBirthWeight

Percent low birth weight

LowBirthWeightPctl

Low birth weight percentile

CardiovascularDisease

Age-adjusted rate of emergency department visits for heart attacks per 10,000

CardiovascularDiseasePctl

Cardiovascular disease percentile

TotalPop

2019 American Community Survey population estimates in census tracts

ChildrenPercLess10

2019 ACS population estimates of the percent per census tract of children under 10 years old

PopPerc10to64

2019 ACS population estimates of the percent per census tract of people between 10 and 64 years old

ElderlyMore64

2019 ACS population estimates of the percent per census tract of elderly 65 years and older

HispanicPerc

2019 ACS population estimates of the percent per census tract of those who identify as Hispanic or Latino

WhitePerc

2019 ACS population estimates of the percent per census tract of those who identify as non-Hispanic white

AfAmericanPerc

2019 ACS population estimates of the percent per census tract of those who identify as non-Hispanic African American or black

NativeAmericanPerc

2019 ACS population estimates of the percent per census tract of those who identify as non-Hispanic Native American

AsianAmericanPerc

2019 ACS population estimates of the percent per census tract of those who identify as non-Hispanic Asian or Pacific Islander

OtherMultiplePerc

2019 ACS population estimates of the percent per census tract of those who identify as non-Hispanic "other" or as multiple races

PopChar

Average of percentiles from the Population Characteristics indicators

PopCharScore

Population Characteristics variable scaled with a range of 0-10. (Used to calculate CES 4.0 Score)

PopCharPctl

Population characteristics percentile

Education

Percent of population over 25 with less than a high school education

EducationPctl

Education percentile

LinguisticIsol

Percent limited English speaking households

LinguisticIsolPctl

Linguistic isolation percentile

Poverty

Percent of population living below two times the federal poverty level

PovertyPctl

Poverty percentile

Unemployment

Percent of the population over the age of 16 that is unemployed and eligible for the labor force

UnemploymentPctl

Unemployment percentile

HousingBurden

Percent housing-burdened low-income households

HousingBurdenPctl

Housing burden percentile

Source

Created by Elizabeth Humphries using data downloaded from the CalEnviroScreen website

Examples

data(calenviroscreen)